datalogger-agent/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.0
structlog==23.2.0
pytest==7.4.3
pytest-mock==3.11.1
pytest-asyncio==0.21.1
